Enhance yt-dlp build, add PDF utility, and update project files
- Improves the `yt-dlp` build script by integrating `uv` for faster dependency management and updates CFLAGS for bootloader compilation.
- Adds a new PowerShell script to batch rename PDF files, targeting those processed by PDF24.
- Includes new AI prompt engineering examples.
- Updates `.gitignore` to exclude `.ttf` font files.

+++ yt-dlp/autobuild-yt-dlp.cmd
@@ -1,10 +1,19 @@
+set USE_UV=1
+
 rem Clone the respective repositories
 git clone https://github.com/pyinstaller/pyinstaller --depth=50
 git clone https://github.com/yt-dlp/yt-dlp --depth=1
-pip cache purge
 
-virtualenv venv --clear
-venv\scripts\activate
+rem Clear cache
+if %USE_UV%==1 (
+    uv cache prune
+    uv venv
+) else (
+    pip cache purge
+    virtualenv .venv --clear
+)
+
+.venv\scripts\activate
 cd pyinstaller/bootloader
 
 rem Cmd and Powershell command
@@ -12,11 +21,17 @@ set PATH=%PATH%;E:\Personal Data\Repositories\mingw64\bin
 rem $env:PATH += 'E:\Personal Data\Repositories\mingw64\bin'
 
 rem Build PyInstaller bootloader using custom gcc binaries
-set CFLAGS=-flto
-set LDFLAGS=-flto
+set CFLAGS=-flto -march=haswell -mtune=generic -pipe
 python .\waf all --gcc -j 21
 cd ..\
-pip3 install -e .
+
+rem Install pyinstaller
+if %USE_UV%==1 (
+    uv pip install -e .
+) else (
+    pip3 install -e .
+)
+
 cd ..\
 cd yt-dlp
 
@@ -27,9 +42,15 @@ python versioninfo.py
 
 rem Build yt-dlp
 set PYTHONOPTIMIZE=2
-pip install -e .
+if %USE_UV%==1 (
+    uv pip install -e .
+) else (
+    pip install -e .
+)
+
 pyinstaller -y --clean yt_dlp/__main__.py --name yt-dlp --noupx --version-file versioninfo.txt --collect-submodules yt-dlp.yt_dlp.utils --icon ../banner.ico --recursive-copy-metadata yt-dlp --recursive-copy-metadata pyinstaller
-pyinstaller -y --onefile --clean yt_dlp/__main__.py --name yt-dlp --noupx --version-file versioninfo.txt --collect-submodules yt-dlp.yt_dlp.utils --icon ../banner.ico --recursive-copy-metadata yt-dlp --recursive-copy-metadata pyinstaller
+pyinstaller -y --onefile yt_dlp/__main__.py --name yt-dlp --noupx --version-file versioninfo.txt --collect-submodules yt-dlp.yt_dlp.utils --icon ../banner.ico --recursive-copy-metadata yt-dlp --recursive-copy-metadata pyinstaller
+
 cp ..\ffmpeg.exe dist\yt-dlp\
 cp -r dist\yt-dlp\ ..\yt_dlp\
 cp README.md ..\yt_dlp\
